VPD-1775 Split BuildReturnSubmissionService from SubmitReturnService
This focuses the business logic in BuildReturnSubmissionService which can be tested more directly, and leaves the job of fetching data and dealing with failures to SubmitReturnService.
