+++ app/services/returns/BuildReturnSubmissionService.scala
@@ -1,0 +1,183 @@
+/*
+ * Copyright 2026 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package services.returns
+
+import config.FrontendAppConfig
+import models.identifiers.{PeriodKey, VpdId}
+import models.obligations.ObligationDetails
+import models.returns.*
+import models.returns.submit.ReturnCreateRequest
+import models.returns.view.*
+import pages.returns.*
+
+import javax.inject.Inject
+import scala.concurrent.ExecutionContext
+
+class BuildReturnSubmissionService @Inject()(
+  totalDutyDueCalculationService: TotalDutyDueCalculationService,
+  config: FrontendAppConfig
+)(using ExecutionContext) {
+
+  private val ZERO_VALUE = BigDecimal("0")
+  private val FLAG_NOT_FILLED = "0"
+  private val FLAG_FILLED = "1"
+
+  def buildSubmission(ua: ReturnsUserAnswers,
+                      obligation: ObligationDetails,
+                      vpdId: VpdId,
+                      periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]): ReturnCreateRequest = {
+
+    val periodKey = PeriodKey(ua.periodKey)
+
+    val vapingProductsProduced = buildVapingProductsProduced(ua, obligation, periodKeyToDutyRateInPencePerMl)
+    val totalDutyDueVapingProducts = vapingProductsProduced.returns.headOption.map(_.dutyDue).getOrElse(ZERO_VALUE)
+
+    val underDeclaration = buildUnderDeclaration()
+    val overDeclaration = buildOverDeclaration()
+    val otherOptions = buildOtherOptions(ua)
+
+    val spoiltProduct = buildSpoiltProduct(ua, vpdId, periodKeyToDutyRateInPencePerMl)
+
+    val totalDutyDue = totalDutyDueCalculationService.calculate(
+      totalDutyDueVapingProducts,
+      underDeclaration,
+      overDeclaration,
+      spoiltProduct
+    )
+
+    val declaration = ua.get(DeclarationPage).getOrElse(
+      // scalafix:off DisableSyntax.throw
+      throw new IllegalStateException("Declaration details are required for submission")
+    )
+
+    ReturnCreateRequest(
+      periodKey = periodKey.toString,
+      vapingProductsProduced = vapingProductsProduced,
+      underDeclaration = underDeclaration,
+      overDeclaration = overDeclaration,
+      spoiltProduct = spoiltProduct,
+      totalDutyDue = totalDutyDue,
+      otherOptions = otherOptions,
+      declaration = declaration
+    )
+  }
+
+  private def buildVapingProductsProduced(ua: ReturnsUserAnswers,
+                                          obligation: ObligationDetails,
+                                          periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]): VapingProductsProduced = {
+    val dutyDeclared = ua.get(DeclareDutyPage).getOrElse(false)
+    val liquidInMl = ua.get(EnterDutyAmountPage).getOrElse(ZERO_VALUE)
+
+    val dutyRateInPencePerMl: Int = periodKeyToDutyRateInPencePerMl(PeriodKey(obligation.periodKey))
+
+    val liquidInLitres = ConvertToLitres(liquidInMl).toLitres
+
+    val dutyDue = (liquidInMl * (BigDecimal(dutyRateInPencePerMl) / 100)).setScale(2, BigDecimal.RoundingMode.DOWN)
+
+    val dutyRateInPoundsPer10Ml = (BigDecimal(dutyRateInPencePerMl) / 100) * 10
+
+    val vapingProductsProduced = if (dutyDeclared) {
+      VapingProductsProduced(vapingProdManufactured = FLAG_FILLED, returns = Seq(
+        RegularReturn(
+          taxType = config.taxType,
+          dutyRate = dutyRateInPoundsPer10Ml,
+          amountProducedLiquid = liquidInLitres,
+          dutyDue = dutyDue
+        )))
+    } else {
+      VapingProductsProduced(vapingProdManufactured = FLAG_NOT_FILLED, returns = Seq())
+    }
+    vapingProductsProduced
+  }
+
+  private def buildUnderDeclaration(): Option[UnderDeclaration] =
+    Some(UnderDeclaration(
+      underDeclFilled = FLAG_NOT_FILLED,
+      reasonForUnderDecl = None,
+      underDeclarationProducts = None
+    ))
+
+  private def buildOverDeclaration(): Option[OverDeclaration] =
+    Some(OverDeclaration(
+      overDeclFilled = FLAG_NOT_FILLED,
+      reasonForOverDecl = None,
+      overDeclarationProducts = None
+    ))
+
+  private def buildSpoiltProduct(ua: ReturnsUserAnswers, vpdId: VpdId, periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]): Option[SpoiltProduct] = {
+    val declareSpoiltProducts = ua.get(DeclareSpoiltProductsPage)
+    val spoiltVolumes = ua.get(SpoiltVolumeByPeriodPage)
+
+    (declareSpoiltProducts, spoiltVolumes) match {
+      case (Some(spoiltProductsDeclared), Some(volumes)) if spoiltProductsDeclared && volumes.nonEmpty =>
+        val spoiltProducts = volumes.map(buildSpoiltProductItem(_, periodKeyToDutyRateInPencePerMl))
+
+        Some(SpoiltProduct(
+          spoiltProductFilled = FLAG_FILLED,
+          spoiltProducts = Some(spoiltProducts)
+        ))
+
+      case (Some(spoiltProductsDeclared), Some(spoiltVolumesByPeriods)) if !spoiltProductsDeclared && spoiltVolumesByPeriods.isEmpty =>
+        Some(SpoiltProduct(
+          spoiltProductFilled = FLAG_NOT_FILLED,
+          spoiltProducts = None
+        ))
+
+      case _ =>
+        None
+    }
+  }
+
+  private def buildSpoiltProductItem(spoiltVolume: SpoiltVolumeByPeriod, periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]) = {
+    val dutyRateInPencePerMl = periodKeyToDutyRateInPencePerMl(spoiltVolume.periodKey)
+    val dutyRateInPoundsPer10Ml = (BigDecimal(dutyRateInPencePerMl) * 10) / 100
+    val volumeInMl = BigDecimal(spoiltVolume.volume)
+    val volumeInLitres = ConvertToLitres(volumeInMl).toLitres
+    val dutyDue = (volumeInMl * (BigDecimal(dutyRateInPencePerMl) / 100)).setScale(2, BigDecimal.RoundingMode.DOWN)
+
+    SpoiltProductItem(
+      returnPeriodAffected = spoiltVolume.periodKey.toString,
+      taxType = config.taxType,
+      dutyRate = dutyRateInPoundsPer10Ml,
+      amountSpoilt = volumeInLitres,
+      dutyDue = dutyDue
+    )
+  }
+
+  private def buildOtherOptions(ua: ReturnsUserAnswers): Option[OtherOptions] = {
+    val dutySuspenseDeclared = ua.get(DeclareDutySuspensePage).getOrElse(false)
+    val dutySuspenseVolumes = ua.get(EnterDutySuspensePage)
+    
+    (dutySuspenseDeclared, dutySuspenseVolumes) match {
+      case (true, Some(volumes)) =>
+        val volumeReceivedInMl = BigDecimal(volumes.volumeReceived)
+        val volumeMovedInMl = BigDecimal(volumes.volumeMoved)
+        
+        Some(OtherOptions(
+          vapingProductUnderDutySuspense = FLAG_FILLED,
+          volumeMovedFromDutySuspense = Some(ConvertToLitres(volumeReceivedInMl).toLitres),
+          volumeMovedToDutySuspense = Some(ConvertToLitres(volumeMovedInMl).toLitres)
+        ))
+      case _ =>
+        Some(OtherOptions(
+          vapingProductUnderDutySuspense = FLAG_NOT_FILLED,
+          volumeMovedFromDutySuspense = None,
+          volumeMovedToDutySuspense = None
+        ))
+    }
+  }
+}
