+++ app/services/returns/SubmitReturnService.scala
@@ -36,14 +36,9 @@ class SubmitReturnService @Inject()(
   submitReturnConnector: SubmitReturnConnector,
   dutyRateService: DutyRateService,
   obligationService: ObligationService,
-  totalDutyDueCalculationService: TotalDutyDueCalculationService,
+  buildReturnSubmissionService: BuildReturnSubmissionService,
   auditService: AuditService,
-  config: FrontendAppConfig
 )(using ExecutionContext) {
-
-  private val ZERO_VALUE = BigDecimal("0")
-  private val FLAG_NOT_FILLED = "0"
-  private val FLAG_FILLED = "1"
 
   def submit(ua: ReturnsUserAnswers)(implicit request: ReturnsDataRequest[?]): Future[ReturnSubmittedResponse] = {
 
@@ -57,158 +52,13 @@ class SubmitReturnService @Inject()(
         case Some(obl) => Future.successful(obl)
         case None => Future.failed(new IllegalStateException(s"No obligation found for period key: ${ua.periodKey}"))
       }
-      submission = buildSubmission(ua, obligation, request.enrolmentVpdId, periodKeyToDutyRateInPencePerMl)
+      submission = buildReturnSubmissionService.buildSubmission(ua, obligation, request.enrolmentVpdId, periodKeyToDutyRateInPencePerMl)
       result <- submitReturnConnector.submitReturn(submission, request.enrolmentVpdId)
     } yield {
       auditService.auditReturnSubmitted(
         SubmitReturnAuditEvent.buildExplicitAuditEvent(submission, result, request.identifiers, obligations))
  
       result
-    }
-  }
-
-  def buildSubmission(ua: ReturnsUserAnswers,
-                      obligation: ObligationDetails,
-                      vpdId: VpdId,
-                      periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int])(using HeaderCarrier): ReturnCreateRequest = {
-
-    val periodKey = PeriodKey(ua.periodKey)
-
-    val vapingProductsProduced = buildVapingProductsProduced(ua, obligation, periodKeyToDutyRateInPencePerMl)
-    val totalDutyDueVapingProducts = vapingProductsProduced.returns.headOption.map(_.dutyDue).getOrElse(ZERO_VALUE)
-
-    val underDeclaration = buildUnderDeclaration()
-    val overDeclaration = buildOverDeclaration()
-    val otherOptions = buildOtherOptions(ua)
-
-    val spoiltProduct = buildSpoiltProduct(ua, vpdId, periodKeyToDutyRateInPencePerMl)
-
-    val totalDutyDue = totalDutyDueCalculationService.calculate(
-      totalDutyDueVapingProducts,
-      underDeclaration,
-      overDeclaration,
-      spoiltProduct
-    )
-
-    val declaration = ua.get(DeclarationPage).getOrElse(
-      // scalafix:off DisableSyntax.throw
-      throw new IllegalStateException("Declaration details are required for submission")
-    )
-
-    ReturnCreateRequest(
-      periodKey = periodKey.toString,
-      vapingProductsProduced = vapingProductsProduced,
-      underDeclaration = underDeclaration,
-      overDeclaration = overDeclaration,
-      spoiltProduct = spoiltProduct,
-      totalDutyDue = totalDutyDue,
-      otherOptions = otherOptions,
-      declaration = declaration
-    )
-  }
-
-  private def buildVapingProductsProduced(ua: ReturnsUserAnswers,
-                                          obligation: ObligationDetails,
-                                          periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]): VapingProductsProduced = {
-    val dutyDeclared = ua.get(DeclareDutyPage).getOrElse(false)
-    val liquidInMl = ua.get(EnterDutyAmountPage).getOrElse(ZERO_VALUE)
-
-    val dutyRateInPencePerMl: Int = periodKeyToDutyRateInPencePerMl(PeriodKey(obligation.periodKey))
-
-    val liquidInLitres = ConvertToLitres(liquidInMl).toLitres
-
-    val dutyDue = (liquidInMl * (BigDecimal(dutyRateInPencePerMl) / 100)).setScale(2, BigDecimal.RoundingMode.DOWN)
-
-    val dutyRateInPoundsPer10Ml = (BigDecimal(dutyRateInPencePerMl) / 100) * 10
-
-    val vapingProductsProduced = if (dutyDeclared) {
-      VapingProductsProduced(vapingProdManufactured = FLAG_FILLED, returns = Seq(
-        RegularReturn(
-          taxType = config.taxType,
-          dutyRate = dutyRateInPoundsPer10Ml,
-          amountProducedLiquid = liquidInLitres,
-          dutyDue = dutyDue
-        )))
-    } else {
-      VapingProductsProduced(vapingProdManufactured = FLAG_NOT_FILLED, returns = Seq())
-    }
-    vapingProductsProduced
-  }
-
-  private def buildUnderDeclaration(): Option[UnderDeclaration] =
-    Some(UnderDeclaration(
-      underDeclFilled = FLAG_NOT_FILLED,
-      reasonForUnderDecl = None,
-      underDeclarationProducts = None
-    ))
-
-  private def buildOverDeclaration(): Option[OverDeclaration] =
-    Some(OverDeclaration(
-      overDeclFilled = FLAG_NOT_FILLED,
-      reasonForOverDecl = None,
-      overDeclarationProducts = None
-    ))
-
-  private def buildSpoiltProduct(ua: ReturnsUserAnswers, vpdId: VpdId, periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int])(using HeaderCarrier): Option[SpoiltProduct] = {
-    val declareSpoiltProducts = ua.get(DeclareSpoiltProductsPage)
-    val spoiltVolumes = ua.get(SpoiltVolumeByPeriodPage)
-
-    (declareSpoiltProducts, spoiltVolumes) match {
-      case (Some(spoiltProductsDeclared), Some(volumes)) if spoiltProductsDeclared && volumes.nonEmpty =>
-        val spoiltProducts = volumes.map(buildSpoiltProductItem(_, periodKeyToDutyRateInPencePerMl))
-
-        Some(SpoiltProduct(
-          spoiltProductFilled = FLAG_FILLED,
-          spoiltProducts = Some(spoiltProducts)
-        ))
-
-      case (Some(spoiltProductsDeclared), Some(spoiltVolumesByPeriods)) if !spoiltProductsDeclared && spoiltVolumesByPeriods.isEmpty =>
-        Some(SpoiltProduct(
-          spoiltProductFilled = FLAG_NOT_FILLED,
-          spoiltProducts = None
-        ))
-
-      case _ =>
-        None
-    }
-  }
-
-  private def buildSpoiltProductItem(spoiltVolume: SpoiltVolumeByPeriod, periodKeyToDutyRateInPencePerMl: Map[PeriodKey, Int]) = {
-    val dutyRateInPencePerMl = periodKeyToDutyRateInPencePerMl(spoiltVolume.periodKey)
-    val dutyRateInPoundsPer10Ml = (BigDecimal(dutyRateInPencePerMl) * 10) / 100
-    val volumeInMl = BigDecimal(spoiltVolume.volume)
-    val volumeInLitres = ConvertToLitres(volumeInMl).toLitres
-    val dutyDue = (volumeInMl * (BigDecimal(dutyRateInPencePerMl) / 100)).setScale(2, BigDecimal.RoundingMode.DOWN)
-
-    SpoiltProductItem(
-      returnPeriodAffected = spoiltVolume.periodKey.toString,
-      taxType = config.taxType,
-      dutyRate = dutyRateInPoundsPer10Ml,
-      amountSpoilt = volumeInLitres,
-      dutyDue = dutyDue
-    )
-  }
-
-  private def buildOtherOptions(ua: ReturnsUserAnswers): Option[OtherOptions] = {
-    val dutySuspenseDeclared = ua.get(DeclareDutySuspensePage).getOrElse(false)
-    val dutySuspenseVolumes = ua.get(EnterDutySuspensePage)
-    
-    (dutySuspenseDeclared, dutySuspenseVolumes) match {
-      case (true, Some(volumes)) =>
-        val volumeReceivedInMl = BigDecimal(volumes.volumeReceived)
-        val volumeMovedInMl = BigDecimal(volumes.volumeMoved)
-        
-        Some(OtherOptions(
-          vapingProductUnderDutySuspense = FLAG_FILLED,
-          volumeMovedFromDutySuspense = Some(ConvertToLitres(volumeReceivedInMl).toLitres),
-          volumeMovedToDutySuspense = Some(ConvertToLitres(volumeMovedInMl).toLitres)
-        ))
-      case _ =>
-        Some(OtherOptions(
-          vapingProductUnderDutySuspense = FLAG_NOT_FILLED,
-          volumeMovedFromDutySuspense = None,
-          volumeMovedToDutySuspense = None
-        ))
     }
   }
 }
